requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.7
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.5.2
MarkupSafe==2.1.3
platformdirs==3.11.0
Pygments==2.17.2
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.2.0
virtualenv==20.24.6
Werkzeug==3.0.1
